requirements.txt
alembic==1.3.2
Babel==2.7.0
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dominate==2.4.0
elasticsearch==7.5.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
Flask==1.1.1
guess-language-spirit==0.5.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.24.2
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
